% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austRia.R
\docType{package}
\name{austRia}
\alias{austRia}
\alias{austRia-package}
\title{austRia: A package containing data about Austria}
\description{
This package contains data about Austria collected from different sources.
}

